requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
ultralytics>=8.3.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
pillow>=10.2.0
pycocotools>=2.0.7
psutil>=5.9.8
python-pptx>=0.6.23
reportlab>=4.1.0
mrcfile>=1.5.0
tqdm>=4.66.0
opencv-python-headless>=4.10.0.84
mmengine>=0.10.5
mmcv-lite>=2.1.0,<2.2.0
mmdet>=3.3.0
